Fix methodHashTable logging again

+++ runtime/compiler/runtime/IProfiler.hpp
@@ -261,7 +261,7 @@ class TR_IPMethodData
    uint32_t getPCIndex() { return _pcIndex; }
    void     setPCIndex(uint32_t i) { _pcIndex = i; }
    TR_IPMethodData* next;
-   int32_t _startPC; // Custom variable to fix methodHashTable printing; makes sure method is still loaded
+   uintptr_t  _startPC; // Custom variable to fix methodHashTable printing; makes sure method is still loaded
 
    private:
    TR_OpaqueMethodBlock *_method;
@@ -290,7 +290,7 @@ struct TR_IPMethodHashTableEntry
    TR_OpaqueMethodBlock      *_method; // callee
    TR_IPMethodData            _caller; // link list of callers and their weights. Capped at MAX_IPMETHOD_CALLERS
    TR_DummyBucket             _otherBucket;
-   int32_t                   _startPC; // Custom variable to fix methodHashTable printing; makes sure method is still loaded
+   uintptr_t                    _startPC; // Custom variable to fix methodHashTable printing; makes sure method is still loaded
 
    void add(TR_OpaqueMethodBlock *caller, TR_OpaqueMethodBlock *callee, uint32_t pcIndex);
    };
